requirements.txt
fastapi
uvicorn
pandas
openpyxl
pyyaml
psycopg[binary]
chardet
phonenumbers
pycountry